add error types
