+++ errors/errors_test.go
@@ -14,10 +14,13 @@ func TestError(t *testing.T) {
 	err := NewBadRequest("translate", "API_KEY_INVALID", "API key not valid. Please pass a valid API key.")
 	werr := fmt.Errorf("wrap %w", err)
 
-	if errors.Is(err, base) {
+	if errors.Is(err, new(Error)) {
 		t.Errorf("should not be equal: %v", err)
 	}
 	if !errors.Is(werr, err) {
+		t.Errorf("should be equal: %v", err)
+	}
+	if !errors.Is(werr, NewBadRequest("translate", "API_KEY_INVALID", "")) {
 		t.Errorf("should be equal: %v", err)
 	}
 
@@ -26,5 +29,9 @@ func TestError(t *testing.T) {
 	}
 	if !errors.As(err, &br) {
 		t.Errorf("should be matchs: %v", err)
+	}
+
+	if reason := Reason(err); reason != "API_KEY_INVALID" {
+		t.Errorf("got %s want: %s", reason, err)
 	}
 }
